tree refactoring
