+++ webclient/src/app/components/classificator-tree/model/tree.model.ts
@@ -1,4 +1,4 @@
-import {ClassificatorItem} from '../../domain/classificator';
+import {ClassificatorItem} from '../../../domain/classificator';
 
 export class Tree {
   level: number;
